pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "conductor"
version = "0.5.0"
description = "The AI-Conductor's Operating System — session lifecycle, governance runtime, product extractor"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0", "jsonschema>=4.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "hypothesis>=6.0", "mypy>=1.10"]
mcp = ["mcp>=0.1"]

[project.scripts]
conductor = "conductor.cli:main"

[tool.setuptools]
packages = ["conductor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
